+++ app/admin/course_events.rb
@@ -28,12 +28,14 @@ ActiveAdmin.register CourseEvent do
     end
   end
 
+  scope I18n.t('activerecord.attributes.course_event.states.active'), :active, default: true
+  scope I18n.t('activerecord.attributes.course_event.states.closed'), :closed
+  scope I18n.t('ui.users.commons.all')                              , :all
+
   index do
     selectable_column
     id_column
-    if current_admin_user.is_root?
-      column(:organization)
-    end
+    column(:organization) if current_admin_user.is_root?
     column(:category)
     column(:course)
     column(:room)