+++ app/controllers/api/ui/v1/products_controller.rb
@@ -1,10 +1,11 @@
 class Api::Ui::V1::ProductsController < Api::Ui::BaseController
+  before_action :set_organization
   before_action :set_product, except: [:index]
 
   respond_to :json
 
   def index
-    @products = current_organization.products.order('id')
+    @products = @organization.products.order('id')
     # Consente di filtrare tramite https://github.com/activerecord-hackery/ransack
     if params[:q].present?
       @products = @products.ransack(params[:q]).result
@@ -22,7 +23,7 @@ class Api::Ui::V1::ProductsController < Api::Ui::BaseController
   private
 
   def set_product
-    @product = current_organization.products.find(params[:id])
+    @product = @organization.products.find(params[:id])
   end
 
   def product_filter_params