+++ app/javascript/users/components/subscriptions/subscriptions_list.vue
@@ -1,0 +1,22 @@
+<template>
+	<v-list three-line>
+		<SubscriptionCard v-for="subscription in subscriptions" :subscription="subscription"></SubscriptionCard>
+	</v-list>
+</template>
+
+<script>
+	import SubscriptionCard from './subscription_card'
+
+  export default {
+    props: {
+			subscriptions: {
+			  type: Array,
+				required: true
+			}
+    },
+
+		components: {
+      SubscriptionCard,
+		},
+	}
+</script>