+++ app/javascript/users/store/store.js
@@ -6,6 +6,7 @@ import * as layout from './modules/layout'
 import * as profile from './modules/profile'
 import * as course_event from './modules/course_event'
 import * as product from './modules/product'
+import * as subscription from './modules/subscription'
 import * as order from './modules/order'
 
 Vue.use(Vuex)
@@ -20,6 +21,7 @@ export default new Vuex.Store({
     profile,
     course_event,
     product,
+    subscription,
     order,
   },
   state: {},